Fix minor issues mentioned in the PR
- Use specific error rather than `RuntimeException`
- Move topological sort to a new file
- Update cyclic inheritance error message
